// src/types/airtable.generated.ts
export interface AirtableLead {
  id: string;
  createdTime: string;
  fields: {
    [key: string]: any;
    /** formula field (fldFVsIGHZkLvnp5O) */
    ID: string;
    /** dateTime field (fldo1bSuO1yjHXjkL) */
    Data?: string;
    /** multipleRecordLinks field (fldPOa4TYVKbLiJTa) */
    Referenza?: string[];
    /** multipleLookupValues field (fldoCxUXtxrl6J7OJ) */
    'Nome referenza'?: string | string[];
    /** multipleRecordLinks field (fldCoDBMqa36c2jJ2) */
    Fonte?: string[];
    /** singleLineText field (fldKeDrS1XiJ47Prc) */
    Nome?: string;
    /** singleSelect field (fld42I37Sfnphcmza) */
    Gender?: 'male' | 'female' | 'unknown';
    /** singleLineText field (fldJJO2MU9E1Q0U6I) */
    Telefono?: string;
    /** email field (fldM2RXG7aRwGXLZU) */
    Email?: string;
    /** multilineText field (fldyt6SwpuJfROOd2) */
    Indirizzo?: string;
    /** number field (fldIzZxsJ9oRQIQy8) */
    CAP?: number;
    /** singleLineText field (fldgTqcRN7Ce4hPrk) */
    Città?: string;
    /** multilineText field (fld7ALmK1onyIj1zX) */
    Esigenza?: string;
    /** singleSelect field (fldzDSwT9nKOnbkan) */
    Stato?: 'Nuovo' | 'Contattato' | 'Qualificato' | 'In Negoziazione' | 'Cliente' | 'Perso' | 'Sospeso';
    /** multilineText field (fldoI9OwYbSTJuDmE) */
    Allegati?: string;
    /** multipleRecordLinks field (fldu170X3rbZBaGdj) */
    Assegnatario?: string[];
    /** singleLineText field (fldL4vLMGuXMCs9yA) */
    Ordini?: string;
    /** multipleRecordLinks field (fldkdo2z38CnPc3Uk) */
    Attività?: string[];
    /** multipleRecordLinks field (fld0mxht6l5onjOfT) */
    'From field: Referenza'?: string[];
    /** multipleRecordLinks field (fld4NKnBTOB0g1QYs) */
    Notes?: string[];
    /** singleLineText field (fldoD2h8AYJf3zx23) */
    Conversations?: string;
    /** singleLineText field (fldwRJVejr2Grk4e8) */
    Order?: string;
    /** multipleRecordLinks field (fldZj5OaJQQwyIHRt) */
    Orders?: string[];
  };
}
// ...
export interface AirtableNotes {
  id: string;
  createdTime: string;
  fields: {
    [key: string]: any;
    /** formula field (fldQ9C8FopcgY5olR) */
    ID: string;
    /** multipleRecordLinks field (fldnNg4T43UMokhNB) */
    Lead?: string[];
    /** multipleRecordLinks field (fldV9X9o3K9xNgbka) */
    User?: string[];
    /** multilineText field (fldcN1Bb8fjT46OyO) */
    Content?: string;
    /** singleSelect field (fldEe2eW9PboYAcRS) */
    Type?: 'Riflessione' | 'Promemoria' | 'Follow-up' | 'Info Cliente';
    /** checkbox field (flduQTxMKNqGqe5i8) */
    Pinned?: boolean;
    /** createdTime field (fldY6MDn5uxeCtkBA) */
    CreatedAt?: any;
  };
}